requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
drf-yasg>=1.21
django-filter>=24.2
gunicorn>=22.0

